Separated Filter from Staff Assignment presenter
Filter is now a subcomponent in the filters folder.
Filter is also using the Presenter pattern.

Jira ticket: CAMS-288
