+++ user-interface/src/staff-assignment/screen/staffAssignmentControls.ts
@@ -1,15 +1,12 @@
 import { RefObject } from 'react';
 import { AssignAttorneyModalRef } from '../modal/AssignAttorneyModal';
 import { ModalRefType } from '@/lib/components/uswds/modal/modal-refs';
-import { CamsUserReference } from '@common/cams/users';
-
-export type StaffAssignmentScreenFilter = {
-  assignee: CamsUserReference;
-};
+import { StaffAssignmentFilterRef } from '../filters/StaffAssignmentFilter';
 
 interface StaffAssignmentControls {
   infoModalRef: React.RefObject<ModalRefType>;
   assignmentModalRef: RefObject<AssignAttorneyModalRef>;
+  filterRef: RefObject<StaffAssignmentFilterRef>;
 }
 
 export type { StaffAssignmentControls };
