+++ user-interface/src/staff-assignment/screen/staffAssignmentStore.ts
@@ -1,11 +1,6 @@
-import { CamsUserReference } from '@common/cams/users';
-import { StaffAssignmentScreenFilter } from './staffAssignmentControls';
+import { StaffAssignmentScreenFilter } from '../filters/StaffAssignmentFilter';
 
 interface StaffAssignmentStore {
-  officeAssignees: CamsUserReference[];
-  setOfficeAssignees(val: CamsUserReference[]): void;
-  officeAssigneesError: boolean;
-  setOfficeAssigneesError(val: boolean): void;
   staffAssignmentFilter: StaffAssignmentScreenFilter | undefined;
   setStaffAssignmentFilter(val: StaffAssignmentScreenFilter | undefined): void;
 }
