+++ user-interface/src/staff-assignment/screen/staffAssignmentStoreReact.ts
@@ -1,19 +1,12 @@
 import { useState } from 'react';
-import { CamsUserReference } from '@common/cams/users';
-import { StaffAssignmentScreenFilter } from './staffAssignmentControls';
+import { StaffAssignmentScreenFilter } from '../filters/StaffAssignmentFilter';
 
 export function useStaffAssignmentStoreReact() {
-  const [officeAssignees, setOfficeAssignees] = useState<CamsUserReference[]>([]);
-  const [officeAssigneesError, setOfficeAssigneesError] = useState<boolean>(false);
   const [staffAssignmentFilter, setStaffAssignmentFilter] = useState<
     StaffAssignmentScreenFilter | undefined
   >();
 
   return {
-    officeAssignees,
-    setOfficeAssignees,
-    officeAssigneesError,
-    setOfficeAssigneesError,
     staffAssignmentFilter,
     setStaffAssignmentFilter,
   };
